feat(decisions): overview submit CTA creates a draft proposal directly
- extract useCreateProposal hook from DecisionActionBar (create empty
  draft -> redirect to its edit page, with pending state + error toast)
- overview hero Submit a proposal button now uses the hook instead of
  linking to the current phase view
- DecisionActionBar reuses the shared hook
