+++ apps/app/src/components/decisions/DecisionActionBar.tsx
@@ -1,22 +1,17 @@
 'use client';
 
-import { useFeatureFlag } from '@/hooks/useFeatureFlag';
-import { useUser } from '@/utils/UserProvider';
-import { trpc } from '@op/api/client';
-import { createSBBrowserClient } from '@op/supabase/client';
 import { Button } from '@op/ui/Button';
 import { Dialog, DialogTrigger } from '@op/ui/Dialog';
 import { LoadingSpinner } from '@op/ui/LoadingSpinner';
 import { Modal, ModalBody, ModalHeader } from '@op/ui/Modal';
 import { RichTextViewer } from '@op/ui/RichTextEditor';
-import { toast } from '@op/ui/Toast';
 import he from 'he';
 import { useParams } from 'next/navigation';
-import { useState } from 'react';
 
-import { useRouter, useTranslations } from '@/lib/i18n';
+import { useTranslations } from '@/lib/i18n';
 
 import { getViewerExtensions } from '../RichTextEditor/editorConfig';
+import { useCreateProposal } from './useCreateProposal';
 
 export const DecisionActionBar = ({
   instanceId,
@@ -32,49 +27,10 @@ export const DecisionActionBar = ({
   showSubmitButton?: boolean;
 }) => {
   const t = useTranslations();
-  const { slug } = useParams();
-  const router = useRouter();
-  const { user } = useUser();
-  const [isCreating, setIsCreating] = useState(false);
-  const supabase = createSBBrowserClient();
-  const utils = trpc.useUtils();
-  const anonymousSigninEnabled = useFeatureFlag('anonymous_signin');
+  const { slug } = useParams<{ slug: string }>();
 
-  const createProposalMutation = trpc.decision.createProposal.useMutation();
-
-  const handleCreateProposal = async () => {
-    setIsCreating(true);
-
-    try {
-      // A public (no-session) visitor has no account to attribute the proposal
-      // to, so give them an anonymous session before creating the draft.
-      if (anonymousSigninEnabled && !user) {
-        const { error } = await supabase.auth.signInAnonymously();
-        if (error) {
-          throw error;
-        }
-
-        // The new session isn't reflected in the cached account query, so
-        // refetch it before navigating — the edit page requires a populated
-        // user in context.
-        await utils.account.getMyAccount.invalidate();
-      }
-
-      const proposal = await createProposalMutation.mutateAsync({
-        processInstanceId: instanceId,
-        proposalData: {}, // Empty draft - user will fill in via edit page
-      });
-
-      // Navigate to edit the newly created draft proposal
-      router.push(`/decisions/${slug}/proposal/${proposal.profileId}/edit`);
-    } catch (error) {
-      setIsCreating(false);
-      toast.error({
-        title: t('Failed to create proposal'),
-        message: error instanceof Error ? error.message : undefined,
-      });
-    }
-  };
+  const { createProposal: handleCreateProposal, isCreating } =
+    useCreateProposal({ instanceId, decisionSlug: slug });
 
   return (
     <div className="flex w-full justify-center">
